feat: refactor source chunk reference handling and integrate into knowledge retrieval
