+++ tests/test_knowledge_retrieval.py
@@ -1,5 +1,6 @@
-from pathlib import Path
 import json
+from pathlib import Path
+
 
 from adapters.knowledge_local import (
     RETRIEVER_POOL_MAX_SIZE,
